Closing stderr is masking the test failure cause
On Windows there is a quoting issue, this tries a different quoting
scheme to fix the test failure.

Other platforms are reporting "more than one plan in TAP output" which
is expected because I stopped closing stdout for the prior dev release.

By closing stdout and not stderr, I should get to see any other
exception messages. For example some platforms were failing with exit
code 2, which is a compile error (package not found in ISA?).

Let's see how this performs on CPAN Testers ...